Se agrega pantalla de pacientes y parte del ABM, AltaPaciente y ModificarPaciente.

+++ Dominio/Paciente.cs
@@ -8,7 +8,7 @@ namespace Dominio
 {
     public class Paciente
     {
-        public int IdPaciente { get; set; }
+        public long IdPaciente { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public string DNI { get; set; }
